Add some responsive tweaks to small label sizes
While working on the ai2html example (which rigidly uses the type and colors laid out in the style guide) I realized there is a need for smaller labels at smaller widths. The standard NGUX body type shrinks at a small width, so the smallest of the label sizes will, too. It diminishes, but does not eliminate, the type hierarchy between label6 and label7, but it's necessary to make charts reasonable at smaller sizes.